requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-jalali-date==1.0.1
django-timezone-field==5.0
django-widget-tweaks==1.4.12
gevent==21.12.0
greenlet==1.1.2
jdatetime==4.1.0
kombu==5.2.4
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
